+++ Examples/Minimal/Resources/Operations.php
@@ -30,11 +30,11 @@ class Operations
     public function pets($pets)
     {
         // Autodetected:
-        // @SWG\Api->path is detected as "/operation/pets"  based on the resoursePath and method-name.
-        // @SWG\Operation->nickname are generated based on the method name.
+        // @Api->path is detected as "/operation/pets"  based on the resoursePath and method-name.
+        // @Operation->nickname are generated based on the method name.
         // When a a api has multiple operations the generated nicknames are suffixed with _$index. Example "pets_0" and "pets_1"
-        // A @SWG\Operation is used directly inside the @SWG\Api (Wrapping @SWG\Operation() in @SWG\Operations() is optional)
-        // A @SWG\Parameter is used directly inside the @SWG\Operation (Wrapping @SWG\Parameter() in @SWG\Parameters() is optional)
+        // A @Operation is used directly inside the @Api (Wrapping @Operation() in @Operations() is optional)
+        // A @Parameter is used directly inside the @Operation (Wrapping @Parameter() in @Parameters() is optional)
 
     }
 }